Added Error Handling for Records not starting with SPACE or Asterisk
Added a small change to error handling of files that do not start with a SPACE or ASTERISK. Instead of raising a DBFError and terminating the program, it instead marks it as an erroneous record in the newly created error.log file and continues.

Errors are marked with current date and time, the record in question (identified by the record[0] field) and the record._data[0] that is causing the error.